[refactor] Refactored source files
