+++ src/services/logger.py
@@ -1,0 +1,4 @@
+import logging
+
+
+uvicorn_logger = logging.getLogger("uvicorn.error")
